safari-600.7-branch/LayoutTests/media/track/track-cue-container-rendering-position-expected.txt


The top of the text track container should be in the bottom 30% of the video element.
EVENT(canplaythrough)

EXPECTED (cueDisplayElement.offsetTop > (video.videoHeight * .70) == 'true') OK
END OF TEST
